wip: migrate broken tests to new env variables
